+++ wagtail_localize_rws_languagecloud/__init__.py
@@ -3,5 +3,5 @@ default_app_config = (
 )
 
 
-VERSION = (0, 7, 2)
+VERSION = (0, 8, 0)
 __version__ = ".".join(map(str, VERSION))