+++ CHANGELOG.md
@@ -5,12 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.8.0] - 2022-04-19
+
+### Changed
+
+- `wagtail_localize_rws_languagecloud` must now appear in `INSTALLED_APPS`
+  ahead of `wagtail.admin` and `wagtail_localize`:
+
+  ```py
+  INSTALLED_APPS = [
+      # ...
+      "wagtail_localize_rws_languagecloud",
+      "wagtail.admin",
+      "wagtail_localize",
+  ]
+  ```
 
 ### Added
 
 - Add new `update_translated_pages` management command
 - Display translation status in locale selector
+- Show custom comment for translations imported from LanguageCloud
+- Documentation improvements
 
 ## [0.7.2] - 2022-03-24